+++ src/komtet_kassa_sdk/order.py
@@ -1,5 +1,6 @@
 # coding: utf-8
-from komtet_kassa_sdk.lib.helpers import apply_discount, to_decimal
+from komtet_kassa_sdk.lib.helpers import apply_discount
+
 from . import PaymentMethod, VatRate